+++ README.md
@@ -14,4 +14,4 @@ Everything is in docker to make testing pretty seamless so just run `make`.
 
 ## Future
 
-I want to store a static list of slug names, or make permanent stulgs I want to use rethinkdb to do this. There is a nice ruby port and I think this would be the perfect project to use Ruby.
+I want to store a static list of slug names, or make permanent slugs I want to use rethinkdb to do this. There is a nice ruby port and I think this would be the perfect project to use Ruby.
