+++ config/routes.rb
@@ -2,8 +2,8 @@ Rails.application.routes.draw do
   # For details on the DSL available within this file, see https://guides.rubyonrails.org/routing.html
   root 'application#index'
 
-  get '/url/:id', to: "application#get_slug"
-  match '/url', to: "application#create_update_slug", via: [:post, :put]
-  match '/url/:id', to: "application#delete_slug", via: [:delete]
+  get '/url/:id', to: "link#get_slug"
+  match '/url', to: "link#create_update_slug", via: [:post, :put]
+  match '/url/:id', to: "link#delete_slug", via: [:delete]
 
 end
