+++ docker-compose.yml
@@ -38,7 +38,7 @@ services:
       context: .
     env_file:
       - .env
-    command: bundle exec rdebug-ide --debug --host 0.0.0.0 --port ${APP_DEBUG_PORT} -- rails s -p ${APP_PORT} -b 0.0.0.0
+    command: bundle exec rdebug-ide --debug --host 0.0.0.0 --port ${APP_DEBUG_PORT} -- bin/rails s -p ${APP_PORT} -b 0.0.0.0
     volumes:
       - api:/var/www
     ports:
